Arreglo de Fallos Funcionalidades mayores arregladas.
